sync sTiles source with adv_sTiles: symbolic-only API for INLA/MPI + accumulated fixes
Brings the portable tree current with adv_sTiles (85 source files + stiles.map):
- symbolic-only preprocessing API (sTiles_init_symbolic / set/get_symbolic_only /
  set_symbolic_done_callback) for MPI perm shipping, exported via stiles.map
- STILES_MPI_* env vars whitelisted in the param validator
- 64-bit task offsets, size_t ss_progress alloc casts, single-tile SEGV fix,
  sparse rank clamp, plus the global-params overhaul and later sparse/tile work
Clean build verified; libstiles.so exports all 5 new public symbols.

+++ tools/TileIndexer/TileIndexer.hpp
@@ -300,7 +300,7 @@ using PagedState = State::Paged;
  */
 int countActiveTiles(const int* row_indices,
                      const int* col_indices,
-                     int nnz,
+                     int64_t nnz,
                      int n,
                      int tile_size,
                      Method method,
