+++ tools/compute/pdpotrfsparse.cpp
@@ -87,13 +87,13 @@ void dpotrf_expansion_from_chol_tasks(TiledMatrix *tiledMatrix, stiles_context_t
 
     const auto &tasks = sTiles::get_chol_tasks(tiledMatrix);
     const auto &offsets = sTiles::get_chol_task_offsets(tiledMatrix);
-    const int start = (rank < static_cast<int>(offsets.size())) ? offsets[rank] : 0;
-    const int end = (rank + 1 < static_cast<int>(offsets.size())) ? offsets[rank + 1] : static_cast<int>(tasks.size());
+    const long long start = (rank < static_cast<int>(offsets.size())) ? offsets[rank] : 0;
+    const long long end = (rank + 1 < static_cast<int>(offsets.size())) ? offsets[rank + 1] : static_cast<long long>(tasks.size());
     
     //export_dense_tiled_matrix(tiledMatrix, "fillin/matrix_before_factorization.txt", rank, false, true);
 
     ss_init(num_tiles_per_dim, num_tiles_per_dim, 0);
-    for (int idx = start; idx < end; ++idx) {
+    for (long long idx = start; idx < end; ++idx) {
         const std::array<int,7> &t = tasks[idx];
         const int myroutine = t[0];
         const int m = t[1];
@@ -168,8 +168,8 @@ void dpotrf_expansion_from_chol_tasks11(TiledMatrix *tiledMatrix, stiles_context
 
     const auto &tasks = sTiles::get_chol_tasks(tiledMatrix);
     const auto &offsets = sTiles::get_chol_task_offsets(tiledMatrix);
-    const int start = (rank < static_cast<int>(offsets.size())) ? offsets[rank] : 0;
-    const int end   = (rank + 1 < static_cast<int>(offsets.size())) ? offsets[rank + 1] : static_cast<int>(tasks.size());
+    const long long start = (rank < static_cast<int>(offsets.size())) ? offsets[rank] : 0;
+    const long long end   = (rank + 1 < static_cast<int>(offsets.size())) ? offsets[rank + 1] : static_cast<long long>(tasks.size());
 
     TileMetaCore *tileMeta = tiledMatrix->tileMetaCore;
     SemisparseTileMetaCore *semiMeta = tiledMatrix->semisparseTileMetaCore;
@@ -187,7 +187,7 @@ void dpotrf_expansion_from_chol_tasks11(TiledMatrix *tiledMatrix, stiles_context
     };
 
     ss_init(num_tiles_per_dim, num_tiles_per_dim, 0);
-    for (int idx = start; idx < end; ++idx) {
+    for (long long idx = start; idx < end; ++idx) {
         const std::array<int,7> &t = tasks[idx];
         const int myroutine = t[0];
         const int m = t[1];
@@ -375,13 +375,13 @@ void dpotrf_reduction_from_chol_tasks(TiledMatrix *tiledMatrix, stiles_context_t
 
     const auto &tasks = sTiles::get_chol_tasks(tiledMatrix);
     const auto &offsets = sTiles::get_chol_task_offsets(tiledMatrix);
-    const int start = (rank < static_cast<int>(offsets.size())) ? offsets[rank] : 0;
-    const int end   = (rank + 1 < static_cast<int>(offsets.size())) ? offsets[rank + 1] : static_cast<int>(tasks.size());
+    const long long start = (rank < static_cast<int>(offsets.size())) ? offsets[rank] : 0;
+    const long long end   = (rank + 1 < static_cast<int>(offsets.size())) ? offsets[rank + 1] : static_cast<long long>(tasks.size());
     sTiles::Logger::info("│   [ESMAIL_CHECK] dpotrf_reduction_from_chol_tasks: rank=" + std::to_string(rank) +
                          ", tasks=[" + std::to_string(start) + "," + std::to_string(end) + ")");
 
     ss_init(num_tiles_per_dim, num_tiles_per_dim, 0);
-    for (int idx = start; idx < end; ++idx) {
+    for (long long idx = start; idx < end; ++idx) {
         const std::array<int,7> &t = tasks[idx];
         const int myroutine = t[0];
         const int m = t[1];
@@ -1026,11 +1026,11 @@ void pdpotrf(stiles_context_t *stile) {
 
 //     const auto &tasks = tiledMatrix->chol_tasks;
 //     const auto &offsets = tiledMatrix->chol_task_offsets;
-//     const int start = (rank < static_cast<int>(offsets.size())) ? offsets[rank] : 0;
-//     const int end   = (rank + 1 < static_cast<int>(offsets.size())) ? offsets[rank + 1] : static_cast<int>(tasks.size());
+//     const long long start = (rank < static_cast<int>(offsets.size())) ? offsets[rank] : 0;
+//     const long long end   = (rank + 1 < static_cast<int>(offsets.size())) ? offsets[rank + 1] : static_cast<long long>(tasks.size());
 
 //     ss_init(num_tiles_per_dim, num_tiles_per_dim, 0);
-//     for (int idx = start; idx < end; ++idx) {
+//     for (long long idx = start; idx < end; ++idx) {
 //         const std::array<int,7> &t = tasks[idx];
 //         const int myroutine = t[0];
 //         const int m = t[1];
@@ -1225,11 +1225,11 @@ void pdpotrf(stiles_context_t *stile) {
 
 //     const auto &tasks = tiledMatrix->chol_tasks;
 //     const auto &offsets = tiledMatrix->chol_task_offsets;
-//     const int start = (rank < static_cast<int>(offsets.size())) ? offsets[rank] : 0;
-//     const int end   = (rank + 1 < static_cast<int>(offsets.size())) ? offsets[rank + 1] : static_cast<int>(tasks.size());
+//     const long long start = (rank < static_cast<int>(offsets.size())) ? offsets[rank] : 0;
+//     const long long end   = (rank + 1 < static_cast<int>(offsets.size())) ? offsets[rank + 1] : static_cast<long long>(tasks.size());
 
 //     ss_init(num_tiles_per_dim, num_tiles_per_dim, 0);
-//     for (int idx = start; idx < end; ++idx) {
+//     for (long long idx = start; idx < end; ++idx) {
 //         const std::array<int,7> &t = tasks[idx];
 //         const int myroutine = t[0];
 //         const int m = t[1];
@@ -1563,11 +1563,11 @@ void pdpotrf(stiles_context_t *stile) {
 
 //     const auto &tasks = tiledMatrix->chol_tasks;
 //     const auto &offsets = tiledMatrix->chol_task_offsets;
-//     const int start = (rank < static_cast<int>(offsets.size())) ? offsets[rank] : 0;
-//     const int end   = (rank + 1 < static_cast<int>(offsets.size())) ? offsets[rank + 1] : static_cast<int>(tasks.size());
+//     const long long start = (rank < static_cast<int>(offsets.size())) ? offsets[rank] : 0;
+//     const long long end   = (rank + 1 < static_cast<int>(offsets.size())) ? offsets[rank + 1] : static_cast<long long>(tasks.size());
 
 //     ss_init(num_tiles_per_dim, num_tiles_per_dim, 0);
-//     for (int idx = start; idx < end; ++idx) {
+//     for (long long idx = start; idx < end; ++idx) {
 //         const std::array<int,7> &t = tasks[idx];
 //         const int myroutine = t[0];
 //         const int m = t[1];
